structure to store logs in file created , json payload error resolved , now entity with relationship are working fine

+++ src/main/java/com/managementportal/ems/Repository/LeaveRepo.java
@@ -3,5 +3,5 @@ package com.managementportal.ems.Repository;
 import com.managementportal.ems.entity.LeavesTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LeaveRepo extends JpaRepository<LeavesTable, Long> {
+public interface LeaveRepo extends JpaRepository<LeavesTable, Integer> {
 }
